Revert "Move benchmark to XCCDF version directory"
This reverts commit 009817418933bfb6fe2cdba7d34b9e58bf374b55.

SCAP is huge. It maybe actually easier if we flatten the parsers and for
a moment forget there are versions of each sub-standard. I believe this
simplification, whilst incorrect, will prove beneficial for early steps.

I believe the versioning can be added later on, and I think adding
versions later on may actually lower the overall cost till completion.

Please hold me accountable to the above assertion.
