+++ pkg/scap/models/xccdf/benchmark_test.go
@@ -8,7 +8,7 @@ import (
 )
 
 func TestRFCFeedParsing(t *testing.T) {
-	doc, err := scap_document.ReadDocumentFromFile("../../../../../examples/scap/xccdf/1.2/test_xccdf_complex_check_nand.xccdf.xml")
+	doc, err := scap_document.ReadDocumentFromFile("../../../../examples/scap/xccdf/1.2/test_xccdf_complex_check_nand.xccdf.xml")
 	assert.Nil(t, err)
 	assert.NotNil(t, doc.Benchmark)
 	assert.Equal(t, doc.Benchmark.XMLName.Space, "http://checklists.nist.gov/xccdf/1.2")
